Factor out code duplication in createAccount
